notion-recipe-generator/requirements.txt
requests==2.32.3
fire==0.6.0
python-dotenv==1.0.1